requirements.txt
dotenv==0.9.9
requests==2.32.3
openpyxl==3.1.5
aiohttp==3.11.16
aiofiles==24.1.0
xmindparser==1.0.9
ijson==3.3.0
olefile==0.47
pillow==11.2.1
pyyaml==6.0.2
python-docx==1.1.2